# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/api/hoteles": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Obtener lista de hoteles paginada con filtros dinámicos",
                "operationId": "f883714cfc557ea53a59a7931770de98",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre del hotel para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direccion",
                        "in": "query",
                        "description": "Dirección del hotel para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "telefono",
                        "in": "query",
                        "description": "Teléfono del hotel para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email del hotel para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sitioWeb",
                        "in": "query",
                        "description": "Sitio web del hotel para filtrar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includeHabitaciones",
                        "in": "query",
                        "description": "Inluye el numero de habitaciones",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hoteles filtrada y paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Crear un nuevo hotel",
                "operationId": "f43481e3fac5bffd59f078e5f009eefa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel creado correctamente"
                    }
                }
            }
        },
...
        "/api/hoteles/{hotel}": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Obtener detalles de un hotel",
                "operationId": "168ddbfe325a93e0c7a2f39c24aff9f5",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includeHabitaciones",
                        "in": "query",
                        "description": "Inluye las habitaciones del hotel",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "includeServicios",
                        "in": "query",
                        "description": "Inluye los servicios del hotel",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del hotel"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Actualizar un hotel",
                "operationId": "c60990bed39813bf7053d1fa76cd93d2",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel actualizado correctamente"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Eliminar un hotel",
                "operationId": "de964fda6ff0939eb79972b553dbf132",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel eliminado correctamente"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Actualizar parcialmente un hotel",
                "operationId": "96f256d800a613ac75e2f86f195baefb",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel actualizado correctamente"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreHotelRequest": {
                "title": "Store Hotel Request",
                "description": "Datos requeridos para crear un nuevo hotel",
                "required": [
                    "nombre",
                    "direccion",
                    "telefono",
                    "email",
                    "sitioWeb"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del hotel (mínimo 5 caracteres, máximo 50)",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "direccion": {
                        "description": "Dirección del hotel (mínimo 5 caracteres, máximo 100)",
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 5
                    },
                    "telefono": {
                        "description": "Teléfono del hotel (máximo 20 caracteres)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "description": "Correo electrónico del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "format": "email",
                        "minLength": 5
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "minLength": 5
                    }
                },
                "type": "object"
            },
            "PutHotelRequest": {
                "title": "Put Hotel Request",
                "description": "Datos requeridos para crear un nuevo hotel",
                "required": [
                    "nombre",
                    "direccion",
                    "telefono",
                    "email",
                    "sitioWeb"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del hotel (mínimo 5 caracteres, máximo 50)",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "direccion": {
                        "description": "Dirección del hotel (mínimo 5 caracteres, máximo 100)",
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 5
                    },
                    "telefono": {
                        "description": "Teléfono del hotel (máximo 20 caracteres)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "description": "Correo electrónico del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "format": "email",
                        "minLength": 5
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "minLength": 5
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}